// src/api/generated/models/ProductResponse.ts
export interface ProductResponse {
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    productId?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    productCode?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    productName?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    productCategory?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    productType?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    unitPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    costPrice?: number;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    currency?: string;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    taxRate?: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponse
     */
    stockQuantity?: number;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    isActive?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof ProductResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProductResponse
     */
    updatedDate?: Date;
}